refactor: consolidate form input components by replacing old TextInput/PasswordInput   with react-hook-form versions
