+++ src/views/Signup/components/PasswordFieldWithInfo.tsx
@@ -1,5 +1,5 @@
 import { Info, WarningCircle } from '@phosphor-icons/react';
-import FormPasswordInput from 'components/FormPasswordInput';
+import PasswordInput from 'components/PasswordInput';
 import PasswordStrengthIndicator from '../../../app/shared/components/PasswordStrengthIndicator';
 import { MAX_PASSWORD_LENGTH } from '../../../app/shared/components/ValidPassword';
 import { FieldError, UseFormRegister } from 'react-hook-form';
@@ -28,7 +28,7 @@ const PasswordFieldWithInfo: React.FC<PasswordFieldWithInfoProps> = ({
   return (
     <>
       <label className="space-y-0.5">
-        <FormPasswordInput
+        <PasswordInput
           className={passwordState ? passwordState.tag : ''}
           placeholder={translate('auth.password')}
           label="password"
